Add files via upload
This is the beginning phase of our documentation. I would like us all to read and to add contributions. By Thursday 12pm, we must be done with this document and have an understanding of the core.
From there,  before we move onto Phase 2 we need to create our credits page with brief bios for each contributor on this project, then we will move onto Phase 2 : Developing the User Authentication (coding the foundation). 
Under the section titled "Choosing the Software/Platform Type", we will make our final decision on the framework we are using in development.
Thank you.